Refactoring and resolve tryes
