+++ js/templates/game-view.js
@@ -2,9 +2,9 @@ import {renderElement, appendElement} from '../utils';
 import GameHeaderView from './items/game-header-view';
 import AnswersHistoryView from './items/answers-history';
 import {globalGameData, GameType} from '../data/game-data';
-import TwoOfTwoGameView from './two-of-two-game-view';
-import OneOfOneGameView from './one-of-one-game-view';
-import OneOfThreeGameView from './one-of-three-game-view';
+import TwoOfTwoGameView from './games/two-of-two-game-view';
+import OneOfOneGameView from './games/one-of-one-game-view';
+import OneOfThreeGameView from './games/one-of-three-game-view';
 import Application from '../application';
 import state from '../data/state';
 import timer from './items/timer';
