+++ js/templates/games/one-of-one-game-view.js
@@ -1,8 +1,8 @@
-import AbstractView from '../abstract-view';
-import AnswerButtonsView from './items/answer-buttons';
-import {ImageType} from '../data/game-data';
-import timer from './items/timer';
-import {renderElement, resizeToProperSize} from '../utils';
+import AbstractView from '../../abstract-view';
+import AnswerButtonsView from '../items/answer-buttons';
+import {ImageType} from '../../data/game-data';
+import timer from '../items/timer';
+import {renderElement, renderImage} from '../../utils';
 
 export default class OneOfOneGameView extends AbstractView {
   get element() {
@@ -18,8 +18,7 @@ export default class OneOfOneGameView extends AbstractView {
   getTemplate() {
     const createOption = (item) => {
       const content = renderElement(``, `form`, `game__content game__content--wide`);
-      const properImage = resizeToProperSize(item);
-      properImage.alt = `Option 1`;
+      const properImage = renderImage(item.ProperImg, `Option 1`);
       const answersButtons = new AnswerButtonsView(`question1`).element;
 
       const soloOption = renderElement(``, `div`, `game__option`);
