+++ js/templates/intro-view.js
@@ -10,8 +10,8 @@ export default class IntroView extends AbstractView {
       Application.showGreeting();
     };
   }
-  getTemplate() {
-    return `<section class="intro"><button class="intro__asterisk asterisk" type="button"><span class="visually-hidden">Продолжить</span>*</button>
+  getTemplate(load) {
+    return `<section class="intro"><button class="intro__asterisk asterisk ${load ? `load` : ``}" type="button"><span class="visually-hidden">Продолжить</span>*</button>
     <p class="intro__motto"><sup>*</sup>${templatesData.intro.text}</p></section>`;
   }
   bind() {
