+++ sass/blocks/greeting.scss
@@ -14,6 +14,7 @@
   position: relative;
   width: 780px;
   height: 740px;
+  margin: 0 auto;
 
   transition: background-image 1s ease-in-out;
   background: url("../img/egg_distort.png") no-repeat center;
