+++ sass/blocks/intro.scss
@@ -8,7 +8,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-
+  margin: 0 auto;
   width: 800px;
   height: 740px;
