+++ sass/scaffolding.scss
@@ -131,6 +131,10 @@ a {
   animation: blink 0.5s linear infinite;
 }
 
+.load {
+  animation: load 1.1s infinite linear;
+}
+
 @keyframes blink {
   0% {
     opacity: 1;
